+++ src/settings/JobProfiles/ViewJobProfile/index.js
@@ -1,1 +1,2 @@
 export * from './ViewJobProfile';
+export * from './NoJobProfilePane';